Updated hgssh3.py to work with Python 3.3
Fix included changing "except ValueError, e:" to "except ValueError as e:"
This was verified to work with Python 2.7 and 3.3.1.
Please note that you will need to make sure the mercurial egg is
available for your Python version or you may get an error about not
being able to import mercurial, or other packages.